pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "pokefinder"
version = "0.1.0"
description = "Pokemon TCG deal finder for the Israeli community"
requires-python = ">=3.11"
dependencies = [
    # Web framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "python-multipart>=0.0.9",
    # Telegram bot
    "python-telegram-bot>=21.0",
    # WhatsApp / Twilio
    "twilio>=9.0.0",
    # Database
    "supabase>=2.5.0",
    "asyncpg>=0.29.0",
    # Scheduling
    "apscheduler>=3.10.0",
    # HTTP client
    "httpx>=0.27.0",
    # Payments
    "stripe>=9.0.0",
    # Scraping
    "playwright>=1.44.0",
    "beautifulsoup4>=4.12.0",
    # Matching & dedup
    "rapidfuzz>=3.9.0",
    "imagehash>=4.3.0",
    "Pillow>=10.0.0",
    "haversine>=2.8.0",
    # Config
    "pydantic-settings>=2.3.0",
    "python-dotenv>=1.0.0",
    "anthropic>=0.27.0",
    # Discord
    "discord.py>=2.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-httpx>=0.30.0",
    "ruff>=0.4.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/pokefinder"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
